tests/wpt/web-platform-tests/html/semantics/embedded-content/the-img-element/null-image-source.html

<!doctype html>
<meta charset="utf-8">
<title>Null image source check for src, srcset and picture parent</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<img id="src_id" src="">
<img id="srcset_id" srcset="">
<picture><img id="parent_picture_id"></picture>
<script>
async_test(function(t) {
    img = document.getElementById('src_id');
    img.onerror = t.step_func(function(e) {
        assert_equals(e.type, "error", "null image source check failed");
        t.done();
    });
}, "img with empty src");

async_test(function(t) {
    img = document.getElementById('srcset_id');
    img.onerror = t.unreached_func("empty srcset fires an error");
    t.step_timeout(function() { t.done(); }, 2000);
}, "img with empty srcset");

async_test(function(t) {
    img = document.getElementById('parent_picture_id');
    img.onerror = t.unreached_func("null img with picture parent fires an error");
    t.step_timeout(function() { t.done(); }, 2000);
}, "img with picture parent");
</script>
